move stuff around to avoid duplication
