+++ src/components/common/SingleProfInfo/singleProfInfo.tsx
@@ -9,8 +9,8 @@ import {
 import Link from 'next/link';
 import React, { useState } from 'react';
 
+import type { GenericFetchedData } from '@/modules/GenericFetchedData/GenericFetchedData';
 import type { RMPInterface } from '@/pages/api/ratemyprofessorScraper';
-import type { GenericFetchedData } from '@/pages/dashboard/index';
 
 type Props = {
   rmp: GenericFetchedData<RMPInterface>;
