+++ src/components/overview/ProfessorOverview/professorOverview.tsx
@@ -9,12 +9,13 @@ import fetchWithCache, {
   cacheIndexNebula,
   expireTime,
 } from '@/modules/fetchWithCache/fetchWithCache';
+import type { GenericFetchedData } from '@/modules/GenericFetchedData/GenericFetchedData';
+import type { GradesType } from '@/modules/GradesType/GradesType';
 import {
   type SearchQuery,
   searchQueryLabel,
 } from '@/modules/SearchQuery/SearchQuery';
 import type { RMPInterface } from '@/pages/api/ratemyprofessorScraper';
-import type { GenericFetchedData, GradesType } from '@/pages/dashboard/index';
 
 const fallbackSrc = 'https://profiles.utdallas.edu/img/default.png';
