+++ src/modules/GradesType/GradesType.ts
@@ -1,0 +1,8 @@
+import type { GradesData } from '@/pages/api/grades';
+
+export type GradesType = {
+  gpa: number;
+  total: number;
+  grade_distribution: number[];
+  grades: GradesData;
+};
